# Performance optimizations in Analysis Request creation